LayoutTests/fast/js/kde/scope-expected.txt

PASS f(2) is 22
PASS successfullyParsed is true
